+++ src/DspSignalBus.cpp
@@ -32,7 +32,7 @@ DspSignalBus::~DspSignalBus()
 
 //=================================================================================================
 
-bool DspSignalBus::SetSignal(unsigned short signalIndex, DspSignal const* newSignal)
+bool DspSignalBus::SetSignal(int signalIndex, DspSignal const* newSignal)
 {
     if (signalIndex < _signals.size() && newSignal != NULL)
     {
@@ -48,7 +48,7 @@ bool DspSignalBus::SetSignal(unsigned short signalIndex, DspSignal const* newSig
 
 bool DspSignalBus::SetSignal(std::string const& signalName, DspSignal const* newSignal)
 {
-    unsigned short signalIndex;
+    int signalIndex;
 
     if (FindSignal(signalName, signalIndex) && newSignal != NULL)
     {
@@ -62,7 +62,7 @@ bool DspSignalBus::SetSignal(std::string const& signalName, DspSignal const* new
 
 //-------------------------------------------------------------------------------------------------
 
-DspSignal* DspSignalBus::GetSignal(unsigned short signalIndex)
+DspSignal* DspSignalBus::GetSignal(int signalIndex)
 {
     if (signalIndex < _signals.size())
     {
@@ -78,7 +78,7 @@ DspSignal* DspSignalBus::GetSignal(unsigned short signalIndex)
 
 DspSignal* DspSignalBus::GetSignal(std::string const& signalName)
 {
-    unsigned short signalIndex;
+    int signalIndex;
 
     if (FindSignal(signalName, signalIndex))
     {
@@ -92,14 +92,14 @@ DspSignal* DspSignalBus::GetSignal(std::string const& signalName)
 
 //-------------------------------------------------------------------------------------------------
 
-bool DspSignalBus::FindSignal(std::string const& signalName, unsigned short& returnIndex) const
+bool DspSignalBus::FindSignal(std::string const& signalName, int& returnIndex) const
 {
     if (signalName == "")
     {
         return false;
     }
 
-    for (unsigned short i = 0; i < _signals.size(); i++)
+    for (int i = 0; i < _signals.size(); i++)
     {
         if (_signals[i].GetSignalName() == signalName)
         {
@@ -113,7 +113,7 @@ bool DspSignalBus::FindSignal(std::string const& signalName, unsigned short& ret
 
 //-------------------------------------------------------------------------------------------------
 
-bool DspSignalBus::FindSignal(unsigned short signalIndex, unsigned short& returnIndex) const
+bool DspSignalBus::FindSignal(int signalIndex, int& returnIndex) const
 {
     if (signalIndex < _signals.size())
     {
@@ -126,14 +126,14 @@ bool DspSignalBus::FindSignal(unsigned short signalIndex, unsigned short& return
 
 //-------------------------------------------------------------------------------------------------
 
-unsigned short DspSignalBus::GetSignalCount() const
+int DspSignalBus::GetSignalCount() const
 {
     return _signals.size();
 }
 
 //-------------------------------------------------------------------------------------------------
 
-void DspSignalBus::ClearValue(unsigned short signalIndex)
+void DspSignalBus::ClearValue(int signalIndex)
 {
     if (signalIndex < _signals.size())
     {
@@ -145,7 +145,7 @@ void DspSignalBus::ClearValue(unsigned short signalIndex)
 
 void DspSignalBus::ClearValue(std::string const& signalName)
 {
-    unsigned short signalIndex;
+    int signalIndex;
 
     if (FindSignal(signalName, signalIndex))
     {
@@ -157,7 +157,7 @@ void DspSignalBus::ClearValue(std::string const& signalName)
 
 void DspSignalBus::ClearAllValues()
 {
-    for (unsigned short i = 0; i < _signals.size(); i++)
+    for (int i = 0; i < _signals.size(); i++)
     {
         _signals[i].ClearValue();
     }
@@ -169,7 +169,7 @@ bool DspSignalBus::_AddSignal(std::string const& signalName)
 {
     if (signalName != "")
     {
-        unsigned short signalIndex;
+        int signalIndex;
         if (FindSignal(signalName, signalIndex))  // if the name already exists
         {
             return false;
